backend/requirements.txt
Django==3.2.21
djangorestframework==3.14.0
django-colorfield==0.10.1
django-filter==22.1
django-import-export==3.2.0
django-modeladmin-reorder==0.3.1
djoser==2.1.0
drf-base64==2.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2_binary==2.9.3
python-dotenv==0.21.1
reportlab==4.0.4
webcolors==1.13